requirements.txt
# LangGraph and LangChain core packages
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# API and HTTP requests
requests>=2.31.0
httpx>=0.27.0

# Environment and configuration
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0

# Logging and utilities
pydantic>=2.8.0
typing-extensions>=4.8.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# KIS API specific dependencies (from GitHub samples)
pycryptodome>=3.20.0  # For token encryption/decryption
websockets>=12.0      # For real-time data streaming

# Optional: For Korean text processing (if needed for news analysis)
# konlpy>=0.6.0

# Optional: For additional LLM providers
# langchain-anthropic>=0.1.0
# langchain-google-genai>=1.0.0